server/requirement.txt
# Web Framework
fastapi
uvicorn
python-multipart

# LangChain & Ecosystem
langchain
langchain-community
langchain-core
langchain-groq

# Vectorstore
# chromadb
pinecone

# Embeddings
# sentence-transformers
# langchain-google-generative-ai
langchain-google-genai

# PDF Parsing
PyPDF  # backend for PyPDFLoader via langchain_community

# Environment Variables
python-dotenv

# Typing & Utilities
pydantic
requests

# Logging (optional but recommended)
loguru

# File Uploads
aiofiles

# Process
tqdm



